Exception handlers

+++ neoProject/src/week4/src/main/java/com/example/week4/service/ProductService.java
@@ -1,27 +1,27 @@
 package com.example.week4.service;
 
 import com.example.week4.entity.Product;
+import com.example.week4.exception.ProductNotFoundException;
 import com.example.week4.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.ArrayList;
-import java.util.NoSuchElementException;
+import java.util.List;
 
 @Service
 public class ProductService {
     private final ProductRepository repository;
 
-    public ProductService( ProductRepository repository) {
+    public ProductService(ProductRepository repository) {
         this.repository = repository;
     }
 
-    public List<Product> getAllProduct(){
+    public List<Product> getAllProduct() {
         return repository.findAll();
     }
     public Product getOneProduct(Long id){
         return repository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("Could not find information about " + id));
+                .orElseThrow(() -> new ProductNotFoundException(id));
     }
     public List<Product> getProductsPriceGreaterThan(double price){
         return repository.findByPriceGreaterThan(price);
@@ -33,7 +33,7 @@ public class ProductService {
         return repository.save(newProduct);
     }
 
-    public Product updateProduct( Product newProduct,Long id){
+    public Product updateProduct(Product newProduct,Long id){
         return repository.findById(id)
                 .map(product -> {
                     product.setName(newProduct.getName());
@@ -41,9 +41,9 @@ public class ProductService {
                     product.setDescription(newProduct.getDescription());
                     return repository.save(product);
                 })
-                .orElseGet(()-> repository.save(newProduct));
+                .orElseThrow(()-> new ProductNotFoundException(id));
     }
-    public List<Product> updateProducts(List<Long> ids, List<Product> updatedProducts){
+    public List<Product> updateProducts(List<Long> ids, List< Product> updatedProducts){
         List<Product> updatedProductsList = new ArrayList<>();
         for(int i = 0; i < ids.size();i++){
             Long id = ids.get(i);
@@ -52,8 +52,12 @@ public class ProductService {
         }
         return updatedProductsList;
     }
-    public void deleteOldProduct( Long id){
-        repository.deleteById(id);
+    public void deleteOldProduct(Long id){
+        if(repository.existsById(id)) {
+            repository.deleteById(id);
+        }else{
+            throw new ProductNotFoundException(id);
+        }
     }
     public void deleteAllProducts() {
         repository.deleteAll();
